Add JSONArray Test
