+++ android-networking/src/androidTest/java/com/androidnetworking/GetStringApiTest.java
@@ -1,20 +1,18 @@
 /*
+ *    Copyright (C) 2016 Amit Shekhar
+ *    Copyright (C) 2011 Android Open Source Project
  *
- *  *    Copyright (C) 2016 Amit Shekhar
- *  *    Copyright (C) 2011 Android Open Source Project
- *  *
- *  *    Licensed under the Apache License, Version 2.0 (the "License");
- *  *    you may not use this file except in compliance with the License.
- *  *    You may obtain a copy of the License at
- *  *
- *  *        http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  *    Unless required by applicable law or agreed to in writing, software
- *  *    distributed under the License is distributed on an "AS IS" BASIS,
- *  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  *    See the License for the specific language governing permissions and
- *  *    limitations under the License.
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
  *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
  */
 
 package com.androidnetworking;
@@ -43,16 +41,12 @@ import okhttp3.mockwebserver.MockWebServer;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 
-/**
- * Created by amitshekhar on 25/03/17.
- */
-
-public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Application> {
+public class GetStringApiTest extends ApplicationTestCase<Application> {
 
     @Rule
     public final MockWebServer server = new MockWebServer();
 
-    public PostStringAndResponseBodyApiTest() {
+    public GetStringApiTest() {
         super(Application.class);
     }
 
@@ -62,16 +56,14 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         createApplication();
     }
 
-    public void testStringPostRequest() throws InterruptedException {
+    public void testStringGetRequest() throws InterruptedException {
 
         server.enqueue(new MockResponse().setBody("data"));
 
         final AtomicReference<String> responseRef = new AtomicReference<>();
         final CountDownLatch latch = new CountDownLatch(1);
 
-        AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
+        AndroidNetworking.get(server.url("/").toString())
                 .build()
                 .getAsString(new StringRequestListener() {
                     @Override
@@ -91,8 +83,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         assertEquals("data", responseRef.get());
     }
 
-
-    public void testStringPostRequest404() throws InterruptedException {
+    public void testStringGetRequest404() throws InterruptedException {
 
         server.enqueue(new MockResponse().setResponseCode(404).setBody("data"));
 
@@ -101,9 +92,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         final AtomicReference<Integer> errorCodeRef = new AtomicReference<>();
         final CountDownLatch latch = new CountDownLatch(1);
 
-        AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
+        AndroidNetworking.get(server.url("/").toString())
                 .build()
                 .getAsString(new StringRequestListener() {
                     @Override
@@ -127,17 +116,15 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         assertEquals("data", errorBodyRef.get());
 
         assertEquals(404, errorCodeRef.get().intValue());
+
     }
 
     @SuppressWarnings("unchecked")
-    public void testSynchronousStringPostRequest() throws InterruptedException {
+    public void testSynchronousStringGetRequest() throws InterruptedException {
 
         server.enqueue(new MockResponse().setBody("data"));
 
-        ANRequest request = AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
-                .build();
+        ANRequest request = AndroidNetworking.get(server.url("/").toString()).build();
 
         ANResponse<String> response = request.executeForString();
 
@@ -145,14 +132,11 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
     }
 
     @SuppressWarnings("unchecked")
-    public void testSynchronousStringPostRequest404() throws InterruptedException {
+    public void testSynchronousStringGetRequest404() throws InterruptedException {
 
         server.enqueue(new MockResponse().setResponseCode(404).setBody("data"));
 
-        ANRequest request = AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
-                .build();
+        ANRequest request = AndroidNetworking.get(server.url("/").toString()).build();
 
         ANResponse<String> response = request.executeForString();
 
@@ -165,16 +149,14 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         assertEquals(404, error.getErrorCode());
     }
 
-    public void testResponseBodyPost() throws InterruptedException {
+    public void testResponseBodyGet() throws InterruptedException {
 
         server.enqueue(new MockResponse().setBody("data"));
 
         final AtomicReference<String> responseRef = new AtomicReference<>();
         final CountDownLatch latch = new CountDownLatch(1);
 
-        AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
+        AndroidNetworking.get(server.url("/").toString())
                 .setExecutor(Executors.newSingleThreadExecutor())
                 .build()
                 .getAsOkHttpResponse(new OkHttpResponseListener() {
@@ -199,7 +181,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         assertEquals("data", responseRef.get());
     }
 
-    public void testResponseBodyPost404() throws InterruptedException {
+    public void testResponseBodyGet404() throws InterruptedException {
 
         server.enqueue(new MockResponse().setResponseCode(404).setBody("data"));
 
@@ -207,9 +189,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         final AtomicReference<Integer> errorCodeRef = new AtomicReference<>();
         final CountDownLatch latch = new CountDownLatch(1);
 
-        AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
+        AndroidNetworking.get(server.url("/").toString())
                 .setExecutor(Executors.newSingleThreadExecutor())
                 .build()
                 .getAsOkHttpResponse(new OkHttpResponseListener() {
@@ -238,14 +218,11 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
     }
 
     @SuppressWarnings("unchecked")
-    public void testSyncResponseBodyPost() throws InterruptedException, IOException {
+    public void testSyncResponseBodyGet() throws InterruptedException, IOException {
 
         server.enqueue(new MockResponse().setBody("data"));
 
-        ANRequest request = AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
-                .build();
+        ANRequest request = AndroidNetworking.get(server.url("/").toString()).build();
 
         ANResponse<Response> response = request.executeForOkHttpResponse();
 
@@ -254,14 +231,11 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
     }
 
     @SuppressWarnings("unchecked")
-    public void testSyncResponseBodyPost404() throws InterruptedException, IOException {
+    public void testSyncResponseBodyGet404() throws InterruptedException, IOException {
 
         server.enqueue(new MockResponse().setResponseCode(404).setBody("data"));
 
-        ANRequest request = AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
-                .build();
+        ANRequest request = AndroidNetworking.get(server.url("/").toString()).build();
 
         ANResponse<Response> response = request.executeForOkHttpResponse();
 
@@ -270,7 +244,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         assertEquals(404, response.getResult().code());
     }
 
-    public void testResponseBodyAndStringPost() throws InterruptedException {
+    public void testResponseBodyAndStringGet() throws InterruptedException {
 
         server.enqueue(new MockResponse().setBody("data"));
 
@@ -278,9 +252,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         final AtomicReference<String> responseStringRef = new AtomicReference<>();
         final CountDownLatch latch = new CountDownLatch(1);
 
-        AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
+        AndroidNetworking.get(server.url("/").toString())
                 .setExecutor(Executors.newSingleThreadExecutor())
                 .build()
                 .getAsOkHttpResponseAndString(new OkHttpResponseAndStringRequestListener() {
@@ -303,7 +275,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         assertEquals("data", responseStringRef.get());
     }
 
-    public void testResponseBodyAndStringPost404() throws InterruptedException {
+    public void testResponseBodyAndStringGet404() throws InterruptedException {
 
         server.enqueue(new MockResponse().setResponseCode(404).setBody("data"));
 
@@ -312,9 +284,7 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
         final AtomicReference<String> errorDetailRef = new AtomicReference<>();
         final CountDownLatch latch = new CountDownLatch(1);
 
-        AndroidNetworking.post(server.url("/").toString())
-                .addBodyParameter("fistName", "Amit")
-                .addBodyParameter("lastName", "Shekhar")
+        AndroidNetworking.get(server.url("/").toString())
                 .setExecutor(Executors.newSingleThreadExecutor())
                 .build()
                 .getAsOkHttpResponseAndString(new OkHttpResponseAndStringRequestListener() {
@@ -340,5 +310,4 @@ public class PostStringAndResponseBodyApiTest extends ApplicationTestCase<Applic
 
         assertEquals(404, errorCodeRef.get().intValue());
     }
-
 }
